+++ src/main/java/de/adorsys/keycloak/config/service/RoleImportService.java
@@ -137,11 +137,19 @@ public class RoleImportService {
                 .findFirst().orElse(null);
 
         if (existingRole != null) {
-            updateClientIfNeeded(realmName, existingRole, roleToImport);
+            updateRoleIfNeeded(realmName, existingRole, roleToImport);
         } else {
-            logger.debug("Create realm-level role '{}' in realm '{}'", roleName, realmName);
-            roleRepository.createRealmRole(realmName, roleToImport);
+            createRole(realmName, roleToImport, roleName);
         }
+    }
+
+    private void createRole(String realmName, RoleRepresentation roleToImport, String roleName) {
+        logger.debug("Create realm-level role '{}' in realm '{}'", roleName, realmName);
+        RoleRepresentation roleToImportWithoutDependencies = CloneUtil.deepClone(
+                roleToImport, RoleRepresentation.class, "composites"
+        );
+
+        roleRepository.createRealmRole(realmName, roleToImportWithoutDependencies);
     }
 
     private void createOrUpdateClientRoles(
@@ -181,12 +189,19 @@ public class RoleImportService {
         if (existingClientRole != null) {
             updateClientRoleIfNecessary(realmName, clientId, existingClientRole, roleToImport);
         } else {
-            logger.debug("Create client-level role '{}' for client '{}' in realm '{}'", roleName, clientId, realmName);
-            roleRepository.createClientRole(realmName, clientId, roleToImport);
+            createClientRole(realmName, clientId, roleToImport, roleName);
         }
     }
 
-    private void updateClientIfNeeded(
+    private void createClientRole(String realmName, String clientId, RoleRepresentation roleToImport, String roleName) {
+        logger.debug("Create client-level role '{}' for client '{}' in realm '{}'", roleName, clientId, realmName);
+        RoleRepresentation roleToImportWithoutDependencies = CloneUtil.deepClone(
+                roleToImport, RoleRepresentation.class, "composites"
+        );
+        roleRepository.createClientRole(realmName, clientId, roleToImportWithoutDependencies);
+    }
+
+    private void updateRoleIfNeeded(
             String realmName,
             RoleRepresentation existingRole,
             RoleRepresentation roleToImport